Do first commands refactoring
