+++ lib/msf/core/exploit/smb/server/share/command.rb
@@ -1,0 +1,15 @@
+# -*- coding: binary -*-
+
+module Msf
+  module Exploit::Remote::SMB::Server
+    module Share
+      module Command
+        require 'msf/core/exploit/smb/server/share/command/close'
+        require 'msf/core/exploit/smb/server/share/command/negotiate'
+        require 'msf/core/exploit/smb/server/share/command/nt_create_andx'
+        require 'msf/core/exploit/smb/server/share/command/read_andx'
+        require 'msf/core/exploit/smb/server/share/command/session_setup_andx'
+      end
+    end
+  end
+end
