+++ lib/msf/core/exploit/smb/server/share/command/close.rb
@@ -1,0 +1,30 @@
+# -*- coding: binary -*-
+
+module Msf
+  module Exploit::Remote::SMB::Server
+    module Share
+      module Command
+        module Close
+          #
+          # Responds to a client CLOSE request
+          #
+          def smb_cmd_close(c, buff)
+            dprint("[SMB_CMD_CLOSE]")
+            pkt = CONST::SMB_CLOSE_PKT.make_struct
+            pkt.from_s(buff)
+
+            pkt = CONST::SMB_CLOSE_RES_PKT.make_struct
+            smb_set_defaults(c, pkt)
+
+            pkt['Payload']['SMB'].v['Command'] = CONST::SMB_COM_CLOSE
+            pkt['Payload']['SMB'].v['Flags1'] = 0x88
+            pkt['Payload']['SMB'].v['Flags2'] = FLAGS2
+            pkt['Payload']['SMB'].v['WordCount'] = 0
+
+            c.put(pkt.to_s)
+          end
+        end
+      end
+    end
+  end
+end
